+++ app/controllers/categories_controller.rb
@@ -76,6 +76,6 @@ class CategoriesController < ApplicationController
 
     # Never trust parameters from the scary internet, only allow the white list through.
     def category_params
-      params.require(:category).permit(:name, :slug, :position)
+      params.require(:category).permit(:top_level, :name, :slug, :position)
     end
 end