+++ config/routes.rb
@@ -44,15 +44,15 @@ Rails.application.routes.draw do
   scope module: 'client_project' do
     get ":id", action: 'root', as: :client_project_root
     get ":id/menu", action: 'menu', as: :client_project_menu
-    get ":id/production/casting", action: 'casting', as: :client_project_casting
-    get ":id/timeline", action: 'timeline', as: :client_project_timeline
+    # get ":id/production/casting", action: 'casting', as: :client_project_casting
+    # get ":id/timeline", action: 'timeline', as: :client_project_timeline
 
-    get ":id/creative/:category_id", action: 'category', as: :client_project_category
-    get ":id/creative/:category_id/:document_id", action: 'document', as: :client_project_category_document
     get ":id/documents/download/:document_id", action: 'download_document', as: :client_project_download_document
+    get ":id/:top_level/:category_id", action: 'category', as: :client_project_category
+    get ":id/:top_level/:category_id/:document_id", action: 'document', as: :client_project_category_document
     get ":id/:category_id", action: 'category'
 
-    get ":id/casting", action: 'casting'
+    # get ":id/casting", action: 'casting'
   end
 
   root 'welcome#index'