+++ db/migrate/20150623232958_add_top_level_to_categories.rb
@@ -1,0 +1,6 @@
+class AddTopLevelToCategories < ActiveRecord::Migration
+  def change
+    add_column :categories, :top_level, :string
+    add_index :categories, :top_level
+  end
+end