+++ test/controllers/client_project_controller_test.rb
@@ -18,43 +18,31 @@ class ClientProjectControllerTest < ActionController::TestCase
   end
 
   test "should get client project concepts category" do
-    get :category, id: projects(:one), category_id: 'concepts'
+    get :category, id: projects(:one), top_level: 'creative', category_id: 'concepts'
     assert_response :success
     assert_not_nil assigns(:project)
   end
 
   test "should get client project treatment category" do
-    get :category, id: projects(:one), category_id: 'treatment'
+    get :category, id: projects(:one), top_level: 'creative', category_id: 'treatment'
     assert_response :success
     assert_not_nil assigns(:project)
   end
 
   test "should get client project script category" do
-    get :category, id: projects(:one), category_id: 'script'
+    get :category, id: projects(:one), top_level: 'creative', category_id: 'script'
     assert_response :success
     assert_not_nil assigns(:project)
   end
 
   test "should get client project boards category" do
-    get :category, id: projects(:one), category_id: 'boards'
+    get :category, id: projects(:one), top_level: 'creative', category_id: 'boards'
     assert_response :success
     assert_not_nil assigns(:project)
   end
 
   test "should get client project references category" do
-    get :category, id: projects(:one), category_id: 'references'
-    assert_response :success
-    assert_not_nil assigns(:project)
-  end
-
-  test "should get client project casting" do
-    get :casting, id: projects(:one)
-    assert_response :success
-    assert_not_nil assigns(:project)
-  end
-
-  test "should get client project timeline" do
-    get :timeline, id: projects(:one)
+    get :category, id: projects(:one), top_level: 'creative', category_id: 'references'
     assert_response :success
     assert_not_nil assigns(:project)
   end